+++ webapp/channels/src/components/user_settings/display/user_settings_theme/user_settings_theme.tsx
@@ -11,7 +11,7 @@ import ExternalLink from 'components/external_link';
 import SettingItemMax from 'components/setting_item_max';
 import SettingItemMin from 'components/setting_item_min';
 import type SettingItemMinComponent from 'components/setting_item_min';
-import ImportThemeModal from 'components/user_settings/import_theme_modal';
+import ImportThemeModal from 'components/user_settings/display/user_settings_theme/import_theme_modal';
 
 import {Constants, ModalIdentifiers} from 'utils/constants';
 import {applyTheme} from 'utils/utils';
