+++ webapp/channels/src/components/user_settings/index.tsx
@@ -28,96 +28,94 @@ export type Props = {
     pluginSettings: {[tabName: string]: PluginConfiguration};
 };
 
-export default class UserSettings extends React.PureComponent<Props> {
-    render() {
-        if (this.props.activeTab === 'profile') {
-            return (
-                <div>
-                    <GeneralTab
-                        user={this.props.user}
-                        activeSection={this.props.activeSection}
-                        updateSection={this.props.updateSection}
-                        updateTab={this.props.updateTab}
-                        closeModal={this.props.closeModal}
-                        collapseModal={this.props.collapseModal}
-                    />
-                </div>
-            );
-        } else if (this.props.activeTab === 'security') {
-            return (
-                <div>
-                    <SecurityTab
-                        user={this.props.user}
-                        activeSection={this.props.activeSection}
-                        updateSection={this.props.updateSection}
-                        closeModal={this.props.closeModal}
-                        collapseModal={this.props.collapseModal}
-                        setRequireConfirm={this.props.setRequireConfirm}
-                    />
-                </div>
-            );
-        } else if (this.props.activeTab === 'notifications') {
-            return (
-                <div>
-                    <NotificationsTab
-                        user={this.props.user}
-                        activeSection={this.props.activeSection}
-                        updateSection={this.props.updateSection}
-                        closeModal={this.props.closeModal}
-                        collapseModal={this.props.collapseModal}
-                    />
-                </div>
-            );
-        } else if (this.props.activeTab === 'display') {
-            return (
-                <div>
-                    <DisplayTab
-                        user={this.props.user}
-                        activeSection={this.props.activeSection}
-                        updateSection={this.props.updateSection}
-                        closeModal={this.props.closeModal}
-                        collapseModal={this.props.collapseModal}
-                        setEnforceFocus={this.props.setEnforceFocus}
-                        setRequireConfirm={this.props.setRequireConfirm}
-                    />
-                </div>
-            );
-        } else if (this.props.activeTab === 'sidebar') {
-            return (
-                <div>
-                    <SidebarTab
-                        activeSection={this.props.activeSection}
-                        updateSection={this.props.updateSection}
-                        closeModal={this.props.closeModal}
-                        collapseModal={this.props.collapseModal}
-                    />
-                </div>
-            );
-        } else if (this.props.activeTab === 'advanced') {
-            return (
-                <div>
-                    <AdvancedTab
-                        activeSection={this.props.activeSection}
-                        updateSection={this.props.updateSection}
-                        closeModal={this.props.closeModal}
-                        collapseModal={this.props.collapseModal}
-                    />
-                </div>
-            );
-        } else if (this.props.activeTab && this.props.pluginSettings[this.props.activeTab]) {
-            return (
-                <div>
-                    <PluginTab
-                        activeSection={this.props.activeSection}
-                        updateSection={this.props.updateSection}
-                        closeModal={this.props.closeModal}
-                        collapseModal={this.props.collapseModal}
-                        settings={this.props.pluginSettings[this.props.activeTab]}
-                    />
-                </div>
-            );
-        }
-
-        return <div/>;
+export default function UserSettings(props: Props) {
+    if (props.activeTab === 'profile') {
+        return (
+            <div>
+                <GeneralTab
+                    user={props.user}
+                    activeSection={props.activeSection}
+                    updateSection={props.updateSection}
+                    updateTab={props.updateTab}
+                    closeModal={props.closeModal}
+                    collapseModal={props.collapseModal}
+                />
+            </div>
+        );
+    } else if (props.activeTab === 'security') {
+        return (
+            <div>
+                <SecurityTab
+                    user={props.user}
+                    activeSection={props.activeSection}
+                    updateSection={props.updateSection}
+                    closeModal={props.closeModal}
+                    collapseModal={props.collapseModal}
+                    setRequireConfirm={props.setRequireConfirm}
+                />
+            </div>
+        );
+    } else if (props.activeTab === 'notifications') {
+        return (
+            <div>
+                <NotificationsTab
+                    user={props.user}
+                    activeSection={props.activeSection}
+                    updateSection={props.updateSection}
+                    closeModal={props.closeModal}
+                    collapseModal={props.collapseModal}
+                />
+            </div>
+        );
+    } else if (props.activeTab === 'display') {
+        return (
+            <div>
+                <DisplayTab
+                    user={props.user}
+                    activeSection={props.activeSection}
+                    updateSection={props.updateSection}
+                    closeModal={props.closeModal}
+                    collapseModal={props.collapseModal}
+                    setEnforceFocus={props.setEnforceFocus}
+                    setRequireConfirm={props.setRequireConfirm}
+                />
+            </div>
+        );
+    } else if (props.activeTab === 'sidebar') {
+        return (
+            <div>
+                <SidebarTab
+                    activeSection={props.activeSection}
+                    updateSection={props.updateSection}
+                    closeModal={props.closeModal}
+                    collapseModal={props.collapseModal}
+                />
+            </div>
+        );
+    } else if (props.activeTab === 'advanced') {
+        return (
+            <div>
+                <AdvancedTab
+                    activeSection={props.activeSection}
+                    updateSection={props.updateSection}
+                    closeModal={props.closeModal}
+                    collapseModal={props.collapseModal}
+                />
+            </div>
+        );
+    } else if (props.activeTab && props.pluginSettings[props.activeTab]) {
+        return (
+            <div>
+                <PluginTab
+                    activeSection={props.activeSection}
+                    updateSection={props.updateSection}
+                    closeModal={props.closeModal}
+                    collapseModal={props.collapseModal}
+                    settings={props.pluginSettings[props.activeTab]}
+                />
+            </div>
+        );
     }
+
+    return null;
 }
